Trapping Rainwater using C
